Extract permissions into separate component
